Libft String.h

+++ ft_memset.c
@@ -3,18 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zelhajou <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: zelhajou <zelhajou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 15:28:07 by zelhajou          #+#    #+#             */
-/*   Updated: 2022/10/20 01:41:58 by zelhajou         ###   ########.fr       */
+/*   Updated: 2022/10/21 22:50:49 by zelhajou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "libft.h"
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int		i;
-	char	*obj;
+	size_t		i;
+	char		*obj;
 
 	i = 0;
 	obj = (char *)s;
